Add pixi features for docs
Adds doc-commands to pixi.
They build and optionally open the documentation
in a browser (only tested on macOS so far.)

Adds pandoc as a conda-dependency.
This would otherwise have to be installed manually
on the host system because it is expected to be
in the path when building the documentation.

stack-info: PR: https://github.com/QSTheory/fftarray/pull/154, branch: SJSeckmeyer/stack/8