Add pytest-split to github CI
This splits the test suite into currently two groups to increase the speed of CI runs further.
The splitting is done on the basis of the measurements in .test_durations.
These will have to be updated manually when the test suite changes in bigger ways.
If some tests change or get added or removed it just makes the split a bit more uneven.
The CI command does not set GROUP_ID but just appends it because this was the cleanest way I found to setup the matrix.

stack-info: PR: https://github.com/QSTheory/fftarray/pull/164, branch: SJSeckmeyer/stack/17