Add pytest-xdist and pytest-split
pytest-xdist parallelizes inside a single process which is useful for local install
(from 6s to 4s on my M1 Max for the short run).

pytest-split is not worth the increased usage in CI minutes on github at the moment,
so only the commands are added to the pyproject.toml and the actual github action file
only contains commented out options.
It splits the test suite into currently two groups.
The splitting is done on the basis of the measurements in .test_durations.
This file is currently not added since this command is not used by default.
These will have to be updated manually when the test suite changes in bigger ways.
If some tests change or get added or removed it just makes the split a bit more uneven.
The CI command would not set GROUP_ID but just append it because this was the cleanest way I found to setup the matrix.

stack-info: PR: https://github.com/QSTheory/fftarray/pull/164, branch: SJSeckmeyer/stack/17