Split dev dependencies into check and doc and tweak CI (#158)
This way the separate github actions do have less packages to install and they run faster.
There is still a "dev" environment which combines every dependency.
In order to also support CI in stacked PRs the triggers are adjusted to trigger on any PR.
The names of the jobs are made more clear.
Tests now use a `test` and not a `build` action.