Add pixi matrix test (#159)
The pip-based test is renamed to make the naming more clear.

This runs the full test suite via pixi with cached environments in CPython 3.10, 3.11, 3.12 and 3.13.

The caching requires that `pixi.lock` is added. This therefore close #133.
Adding `pixi.lock` does not influence projects depending on this library and therefore should be fine.