Remove slow tests from pip CI run (#160)
Before this change this is the slowest overall CI job. Since the full test suite is already run via pixi (which is faster with its setup, especially when cached) the pip CI job is only there for testing the more traditional way of installation and does not have to run every test.