Remove pyfftw (#179)
Its installation caused a lot of problems in different configurations and benchmarks showed that either jax or numpy are competitive on a M1 Max, Ryzen 7950X3D and Epyc 7543.
So for simplicity in the next refactoring this commit removes it.
But it is possible to design it into the code at a later point.