Move to Python Array API
This includes multiple changes.
Array arguments are called `x` if there is only one and otherwise by `x1`, `x2`, etc. if there are no more specific names.
Functions which directly pass through to the Python Array API are positional- and kw-only as mandated by the standard.
The `Backend` classes are completely removed.
The `FFTArray` class always carries its namespace in its state, in order to not always recreate that namespace.
The default backend gets replaced by two defaults:
`default_xp`: The default Array API namespace. The value here is always automatically wrapped by array_api_compat which might however pass through the namespace if it is fully compliant.
`default_dtype`: The default name (as a str) of the dtype which is then turned into an actual dtype by getting the attribute of that name from the Array API namespace.
The default can only be a (complex) floating point number since it is used only for coordinate array creation.

`into` gets split into separate setters and getters for `space`, `factors_applied`, `eager` and `xp`.
This makes the order in which the changes are applied clearer and the individual functions simpler.
The fft code is also restructured to require less defensive copy and be clearer to read.
`FFTArray` gains an invariant to only allow `factors_applied=False` for any dimension if the values have a `complex floating` dtype.
The removes a lot of potential inconsistencies and potential confusion which operations are valid and when upcasting happens.
This change also required splitting the general `test_fft_array_lazyness` into an int and a float variant.
This will be revisited later.

The tested Array API implementations are now `numpy`, `jax.numpy` and `array_api_strict`.
The latter is only meant for testing and helps to keep the whole library Array API conform.
New tests are now in a separate test folder where all tests will be moved to in the future.
jax is now in part an optional backend in testing but that transition is not completed yet.

The jax pytrees get moved into a separate function which currently must be called by the user.

Because the Python Array API standard has different rules for `__add__` versus `xp.add` and similar functions the dunder-methods get their own special implementation path.
This allows moving the other element-wise functions into a separate module.
All element-wise functions (except dunder methods) are now also tested and have auto-generated doc-strings referencing the Python Array API.

The formatting code is cleaned up and gets tests since it broke under Array API strict.

This introduces also two new creation functions `coords_array` and `full` which are used already internally in the shift functions.

Co-authored-by: Christian Struckmann <56967696+cstruckmann@users.noreply.github.com>

stack-info: PR: https://github.com/QSTheory/fftarray/pull/185, branch: SJSeckmeyer/stack/28

Add chck_consistency to full