Refactor transform signs computation to use look up tables (#187)
This makes the implementation more straight-forward to read and extend compared to the old more complicated mutation-based logic.

This PR also fixes #189 and a bug in the handling of operations involving scalars.
If a scalar needed a phase-factor applied this was not properly upcasted.
This refactors `test_transform_application` to properly test for #189 and also other operations and adds `test_transform_application_scalar` to test scalar handling.

Co-authored-by: Christian Struckmann <56967696+cstruckmann@users.noreply.github.com>
Co-authored-by: Gabriel Müller <51076825+gabmueller@users.noreply.github.com>