Upgrade Python Array API v2024.12 (#229)
This includes fixes of the tests for the new version of the Array API and `array-api-strict`.

`conj` and `real` accept all integral and real floating types as of v2024.12
`astype` is  more strictly checked in array-api-strict starting from 2.3.
Therefore the test data creation is improved to create more meaningful values without abusing `into_dtype`.
The more diverse test data requires loosening some of the equality checks.