Add command check_all (#227)
The pixi command check_all runs the complete test suite and all the linter and type-checker.
Add a script to run locally all checks on all python versions and building the docs in doc environment.
This enables to run a complete test of everything offline with a single command  before pushing.