Remove default precision
It is impossible to shield the user from the intricate differences in precision
handling and dtype promotion of different array libraries.
So FFTArray is just a as thin as possible wrapper around the underlying library
and does not interfere there at all.

Co-authored-by: Gabriel Müller <g.mueller@iqo.uni-hannover.de>
Co-authored-by: Christian Struckmann <56967696+cstruckmann@users.noreply.github.com>

stack-info: PR: https://github.com/QSTheory/fftarray/pull/241, branch: SJSeckmeyer/stack/4