Replace numpy in all shape and phase factor calculation with pure python lists.
This is in preparation for torch.compile which would try to move
the NumPy calculations from trace time to run time.
Additionally it had problems to parse the TwoOperandTransforms class therefore it is replaced
by a tuple of lists.

Co-authored-by: Gabriel Müller <51076825+gabmueller@users.noreply.github.com>

stack-info: PR: https://github.com/QSTheory/fftarray/pull/243, branch: SJSeckmeyer/stack/6