Restructure creation functions (#245)
It is impossible to shield the user from the intricate differences in precision
handling and dtype promotion of different array libraries.
So FFTArray is just a as thin as possible wrapper around the underlying library
and does not interfere there at all.
Therefore dtype and device are always just passed through to the underlying library
and the libaries behavior is predictably replicated.

This also fixes the tests of fa.full to follow the behavior of the underlying library
instead of enforcing the specific behavior of keeping the dtype
of a passed in array scalar.

Co-authored-by: Gabriel Müller <51076825+gabmueller@users.noreply.github.com>
Co-authored-by: Christian Struckmann <56967696+cstruckmann@users.noreply.github.com>