Workaround for cross-library conversion (#248)
Since it does not accept arbitrary Python Array API implementations this should
be fully conformant and not suddenly break.

This patch comes before pytorch because then we can keep the xp conversion tests as is.

I am not decided on where to best put this this function and whether to expose it externally.
There are some potential performance improvements for it in the future because for example cupy and pytorch
can exchange CUDA arrays.
But for simplicity this optimization is not done here.