Add pytorch (#249)
This removes some smaller non-standard compliant behavior from the tests.
Additionally it adds a version of the split_step test which tests torch.compile.
It also remove unsigned integers from some tests because pytorch does not support.
Although those are part of the standard this is okay because these tests should test FFTArray
and not stanard-compliance of other libraries.
torch seems to choke on some shared lock when starting 64 test runners, so for pytest needs to run single-threaded.

Co-authored-by: Christian Struckmann <56967696+cstruckmann@users.noreply.github.com>
Co-authored-by: Gabriel Müller <51076825+gabmueller@users.noreply.github.com>