Fix pip_test on github (#257)
Enable the pip_test on each pull request.
Install torch 2.6.0 manually before running the tests.
Version 2.4.0 of array-api-strict also broke a few things, so fix it to 2.3.1 in order to get a minimum working version.
Jax 0.6.2 also breaks a test, fix it to the 0.5.3 from the pixi.lock for now and it will be updated in a later PR.