Update array-api-strict to 2.4.0 and fix tests. (#258)
array-api-strict 2.4.0 now has as the minimum Python version 3.12.
Therefore the tests for lower versions in CI are deactivated for now.
To make sure that all these changes are properly tested before hitting main, enable the minimal CI tests for every python version in every PR.
Also reenable the pip_test.yml after a push to main.

Remove the test of combining 0d arrays with fftarray Arrays.
This behavior is not guaranteed by the standard and flagged by array-api-strict v2.4.
 https://github.com/data-apis/array-api/issues/399
The Gaussians and Bragg Beam Splitter both need some little tweaks to always use either a Python Scalar or an fftarray Array instances which are then backed by the tested backend which might be array_api_strict.