Update jax (#260)
Jax changed the exception type when a tracer gets injected into a metadata field which broke the `test_dynamically_traced_coords` test.
Since jax does not guarantee whether and if yes which exception thrown and we should only test our own exceptions, this change removes all changes related to this.