Adjust minimum jax and pytorch versions
JAX officially supports the Array API standard from 0.4.32 onwards, but that release was yanked, so list 0.4.33.
We do not have enough experionce with PyTorch to really list a minimum version.
In principle it is the responsibility of the user to supply an array library which supports the Python Array API standard sufficiently well.