Move scipy out of helpers dependency group and rename it to z3 (#278)
scipy is only used in the examples so just move it next to bokeh.
That makes the "helpers" group simpler since it only contains the constraint solver.
That dependency is optional in order to allow the use in more embedded environments like Pyodide and also in the future when the z3 package might not be easy to install anymore.