Fix math rendering on pypi
Remove coloneqq from README.md because it is not supported.
Add an rst version of the README, because only that supports proper math rendering on pypi.
But the README.md is necessary for math rendering on github.
So this also adds a covnersion command and a test that the generated rst matches the current one.
Since the test requires pandoc, it is skippable for the pip installation test.

stack-info: PR: https://github.com/QSTheory/fftarray/pull/302, branch: SJSeckmeyer/stack/1