Move all development dependencies under dependency-groups
These do not enforce any versions without explicit installation anyways.
Since it is in principle the responsibility of the user to provide an Array API compliant library this is also fine.
And if they are in dependency-group they are not publicly announced on pypi or built into the wheel.

Also move in the array library dependencies because they are not enforced if they are not used at installation.
Additionally add twine, to make uploading simpler and group hatch and twine into a group named "deploy".