Convert the two species ground state example into a notebook (#316)
This enables the rendering of it in the documentation with the other examples.
Since this example needs a GPU and cannot be built in CI due to performance requirements, its results are stored in the notebook. Since the ntoebook then becomes pretty large, the notebook is added to git lfs.
To keep the notebook simpler, the figure rendering for the versions in the paper is removed.
This also allows to get rid of the injections of the examples directory into PYTHONPATH and MYPYATH.