Add second pip test that solves dependencies at execution time and not requirements.txt
This allows to have an indicator if something new in the ecosystem would break a fresh install of the tests right now.
This has become simpler with the new dependecy-groups refactor.
The pip_test_new_solve will not be a requirement for merging since this is not a change of the actual code change that is being tested.